Changed FFT API to be integer based. Fixed write-back of float FFTs.

+++ projects/plugin/src/Engine.cpp
@@ -555,7 +555,7 @@ namespace ape
 		auxMatrix.resizeChannels(Signalizer::OscilloscopeContent::NumColourChannels);
 		tempBuffer.resizeChannels(ioConfig.outputs);
 		
-		getOscilloscopeData().setTriggeringChannel(ioConfig.inputs + 1);
+		getOscilloscopeData().setTriggeringChannel(static_cast<int>(ioConfig.inputs + 1));
 
 	}
