+++ projects/plugin/src/PluginState.cpp
@@ -167,7 +167,7 @@ namespace ape
 		else
 		{
 			if(!triggerSetThroughAPI)
-				api::setTriggeringChannel(sharedObject.get(), config.inputs + 1);
+				api::setTriggeringChannel(sharedObject.get(), static_cast<int>(config.inputs + 1));
 
 			for (std::size_t i = 0; i < parameters.size(); ++i)
 			{
